requirements.txt
Django==3.2.10
django-ckeditor==6.2.0
easy-thumbnails==2.8
ipython==7.28.0
Pillow==8.4.0
wheel==0.37.1

dj-database-url==0.5.0
django-storage-url==0.5.0
whitenoise==5.2.0
boto3==1.14.49

psycopg2==2.8.5

uwsgi==2.0.19.1